chore: migrate src/Browser to TS
This was an interestng migration and hit upon some tweaks elsewhere in
the codebase where we can gain stricter typing now more code is ported
to TS.
